Cleaned up the base resource pack manager class
